Adding Spring Security

+++ src/main/java/net/engineeringdigest/journalApp/services/JournalEntryService.java
@@ -6,6 +6,7 @@ import net.engineeringdigest.journalApp.repository.JournalEntryRepository;
 import org.bson.types.ObjectId;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
 
@@ -25,12 +26,23 @@ import java.util.Optional;
 
 
 
+    @Transactional
     public void saveEntry (JournalEntry journalEntry, String userName){
-        User user = userService.findByUserName(userName);
-        journalEntry.setDate(LocalDateTime.now());
-        JournalEntry saved = journalEntryRepository.save(journalEntry);
-        user.getJournalEntries().add(saved);
-        userService.saveEntry(user);
+
+        try{
+
+            User user = userService.findByUserName(userName);
+            journalEntry.setDate(LocalDateTime.now());
+            JournalEntry saved = journalEntryRepository.save(journalEntry);
+            user.getJournalEntries().add(saved);
+            userService.saveUser(user);
+
+        }
+        catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException("Entry cannot be saved");
+        }
+
     }
     public void saveEntry (JournalEntry journalEntry){
        journalEntryRepository.save(journalEntry);
@@ -51,11 +63,28 @@ import java.util.Optional;
 
     }
 
-    public void deleteById(ObjectId id, String userName){
+    @Transactional
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removed = false;
+        try{
+            User user = userService.findByUserName(userName);
+             removed = user.getJournalEntries().removeIf(x -> x.getId().equals(id));
+            if(removed){
+                userService.saveUser(user);
+                journalEntryRepository.deleteById(id);
+            }
+        }
+        catch(Exception e){
+            System.out.println(e);
+            throw new RuntimeException("An error occured while deleting the entry.",e);
+        }
 
-        User user = userService.findByUserName(userName);
-        user.getJournalEntries().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepository.deleteById(id);
+        return removed;
+
+
     }
+
+//    public List<JournalEntry> findByUserName(String userName){
+//
+//    }
 }
